+++ app/components/DesktopActionsButtons/messages.js
@@ -1,0 +1,21 @@
+/*
+ * Copyright © 2015-2017 Serenova, LLC. All rights reserved.
+ */
+
+/*
+ * DesktopActionsButtons Messages
+ *
+ * This contains all the text for the DesktopActionsButtons component.
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  cancel: {
+    id: 'app.containers.DesktopActionsButtons.CancelButton',
+    defaultMessage: 'Cancel',
+  },
+  transfer: {
+    id: 'app.containers.DesktopActionsButtons.TransferButton',
+    defaultMessage: 'Transfer',
+  },
+});