+++ app/containers/ActionsMenu/messages.js
@@ -1,0 +1,25 @@
+/*
+ * Copyright © 2015-2017 Serenova, LLC. All rights reserved.
+ */
+
+/*
+ * ActionsMenu Messages
+ *
+ * This contains all the text for the ActionsMenu component.
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  actions: {
+    id: 'app.containers.ActionsMenu.actionsButton',
+    defaultMessage: 'Actions',
+  },
+  cancel: {
+    id: 'app.containers.ActionsMenu.CancelButton',
+    defaultMessage: 'Cancel',
+  },
+  transfer: {
+    id: 'app.containers.ActionsMenu.TransferButton',
+    defaultMessage: 'Transfer',
+  },
+});