+++ app/containers/TransferMenu/selectors.js
@@ -8,12 +8,17 @@ const selectAgentDesktopDomain = (state) => state.get('agentDesktop');
 
 const selectWarmTransfers = createSelector(
   selectAgentDesktopDomain,
-  (agentDesktop) =>
-    agentDesktop
-      .get('interactions')
-      .find((interaction) => interaction.get('channelType') === 'voice')
-      .get('warmTransfers')
-      .toJS()
+  (agentDesktop) => {
+    if (agentDesktop.interactions !== undefined) {
+      return agentDesktop
+        .get('interactions')
+        .find((interaction) => interaction.get('channelType') === 'voice')
+        .get('warmTransfers')
+        .toJS();
+    } else {
+      return null;
+    }
+  }
 );
 
 const selectQueues = createSelector(selectAgentDesktopDomain, (agentDesktop) =>